fixed delete product, need to enable delete it after assignment to transport

+++ lib/core/shared/assign_card_product.dart
@@ -128,20 +128,23 @@ class AssignCardProduct extends StatelessWidget {
               children: <Widget>[
                 ElevatedButton(
                   onPressed: () async {
-                    await db.deleteProductFromProductList(product.id);
-                    await db.deleteProductFromTransportList(
-                        product.id, product.assignedTransportId);
-                    await db.deleteProductFromGivitUserList(product.id);
-                    Transport transport =
-                        await db.getTransportByID(product.assignedTransportId);
-                    if (transport.products.length == 0) {
-                      await db.deleteTransportFromTransportList(transport.id);
-                      await db.updateAssignGivitUsers(transport.carriers, {
-                        "Transports":
-                            FieldValue.arrayRemove(['${transport.id}'])
-                      });
-                    }
                     Navigator.of(context).pop();
+                    await db.deleteProductFromProductList(product.id);
+                    await db.deleteProductFromGivitUserList(product.id);
+                    if (product.assignedTransportId != '') {
+                      Transport transport = await db
+                          .getTransportByID(product.assignedTransportId);
+                      await db.deleteProductFromTransportList(
+                          product.id, product.assignedTransportId);
+
+                      if (transport.products.length == 0) {
+                        await db.deleteTransportFromTransportList(transport.id);
+                        await db.updateAssignGivitUsers(transport.carriers, {
+                          "Transports":
+                              FieldValue.arrayRemove(['${transport.id}'])
+                        });
+                      }
+                    }
                   },
                   child: Text("מחיקה"),
                 ),
